re-separate IrElementTransformer implementation
