+++ backintime-plugin/compiler/src/main/kotlin/com/github/kitakkun/backintime/compiler/backend/BackInTimeIrGenerationExtension.kt
@@ -1,6 +1,9 @@
 package com.github.kitakkun.backintime.compiler.backend
 
-import com.github.kitakkun.backintime.compiler.backend.transformer.BackInTimeDebuggableClassTransformer
+import com.github.kitakkun.backintime.compiler.backend.transformer.capture.BackInTimeDebuggableCaptureLazyDebuggablePropertyAccessTransformer
+import com.github.kitakkun.backintime.compiler.backend.transformer.capture.BackInTimeDebuggableCaptureMethodInvocationTransformer
+import com.github.kitakkun.backintime.compiler.backend.transformer.capture.BackInTimeDebuggableConstructorTransformer
+import com.github.kitakkun.backintime.compiler.backend.transformer.implement.BackInTimeDebuggableImplementTransformer
 import com.github.kitakkun.backintime.compiler.configuration.BackInTimeCompilerConfiguration
 import org.jetbrains.kotlin.backend.common.extensions.IrGenerationExtension
 import org.jetbrains.kotlin.backend.common.extensions.IrPluginContext
@@ -13,7 +16,10 @@ class BackInTimeIrGenerationExtension(
     override fun generate(moduleFragment: IrModuleFragment, pluginContext: IrPluginContext) {
         val context = BackInTimePluginContext(baseContext = pluginContext, config = config, moduleFragment = moduleFragment)
         with(context) {
-            moduleFragment.transformChildrenVoid(BackInTimeDebuggableClassTransformer())
+            moduleFragment.transformChildrenVoid(BackInTimeDebuggableConstructorTransformer())
+            moduleFragment.transformChildrenVoid(BackInTimeDebuggableImplementTransformer())
+            moduleFragment.transformChildrenVoid(BackInTimeDebuggableCaptureMethodInvocationTransformer())
+            moduleFragment.transformChildrenVoid(BackInTimeDebuggableCaptureLazyDebuggablePropertyAccessTransformer())
         }
     }
 }
